refactor(sidebar-routes): split eventRoutes/sidebar.ts by event family
Replace src/eventRoutes/sidebar.ts (577 LOC) with a sidebar/ directory
that groups the 22 handlers by domain:

- resize.ts    - sidebar resize / resize-end (layout column token)
- project.ts   - 6 handlers: remove, toggle-expand, rename,
                 open-finder, copy-path, set-worktree-base
- session.ts   - delete-session + rename-session (consent-aware)
- worktree.ts  - 10 handlers: create, switch, open-in-new-tab,
                 start-session, remove, cancel-pending, retry-pending,
                 rename, open-finder, copy-path
- extension.ts - toggle-extension (bridge command)
- chrome.ts    - handleSectionedSelect (sidebar + model-picker +
                 appearance-menu select dispatch by sectionId)
- index.ts     - barrel re-exporting all 22 handlers

Public surface unchanged. eventRoutes/index.ts imports from
"./sidebar" which now resolves to the directory's index.ts. All 5
event-route precedence layers (shell-consent, extension routes,
built-ins) and the existing sidebar.test.ts suite remain intact.
